# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/internal/simulator/balance-changes": {
      "post": {
        "tags": [
          "balance-change-simulator-endpoint"
        ],
        "operationId": "publish",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "pattern": "Bearer\\s+\\S+",
              "type": "string"
            }
          },
          {
            "name": "TPP-ID",
            "in": "header",
            "required": true,
            "schema": {
              "maxLength": 15,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "Provider-ID",
            "in": "header",
            "required": true,
            "schema": {
              "maxLength": 8,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "Request-ID",
            "in": "header",
            "required": true,
            "schema": {
              "maxLength": 60,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "Request-DateTime",
            "in": "header",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d+)?Z$",
              "type": "string"
            }
          },
          {
            "name": "JWS-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BalanceChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedBalanceEvent"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BalanceChangeRequest": {
        "required": [
          "accountId",
          "amount",
          "balanceAfter",
          "currency",
          "direction",
          "occurredAt",
          "transactionId"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "accountId": {
            "maxLength": 34,
            "minLength": 0,
            "type": "string"
          },
          "transactionId": {
            "maxLength": 70,
            "minLength": 0,
            "type": "string"
          },
          "amount": {
            "minimum": 0.0,
            "exclusiveMinimum": true,
            "type": "number"
          },
          "currency": {
            "minLength": 1,
            "pattern": "^[A-Z]{3}$",
            "type": "string"
          },
          "direction": {
            "type": "string",
            "enum": [
              "CREDIT",
              "DEBIT"
            ]
          },
          "balanceAfter": {
            "type": "number"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AcceptedBalanceEvent": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      }
    }
  }
}